requirements.txt
pywebview==5.1
pycryptodome==3.20.0
pyinstaller==6.9.0
winrt-runtime==2.1.0
winrt-Windows.ApplicationModel==2.1.0
winrt-Windows.Foundation==2.1.0
winrt-Windows.Foundation.Collections==2.1.0
winrt-Windows.Management==2.1.0
winrt-Windows.Management.Deployment==2.1.0
winrt-Windows.Storage==2.1.0